LayoutTests/http/tests/security/isolatedWorld/bypass-main-world-csp-for-xhr-expected.txt

CONSOLE MESSAGE: Refused to connect to http://127.0.0.1:8000/xmlhttprequest/resources/access-control-basic-allow.cgi because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to connect to http://127.0.0.1:8000/xmlhttprequest/resources/access-control-basic-allow.cgi because it does not appear in the connect-src directive of the Content Security Policy.
Tests that isolated worlds can have XHRs that the page's CSP wouldn't allow.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Synchronous XHR same-origin from main world
PASS: XHR.send threw an exception.
Asynchronous XHR same-origin from main world
PASS: XHR.send did not throw an exception.
PASS: XHR.send received an error event.
Synchronous XHR same-origin from isolated world
PASS: XHR.send did not throw an exception.
Asynchronous XHR same-origin from isolated world
PASS: XHR.send did not throw an exception.
PASS: XHR.send received a load event.
Synchronous XHR cross-origin from isolated world
PASS: XHR.send did not throw an exception.
Asynchronous XHR cross-origin from isolated world
PASS: XHR.send did not throw an exception.
PASS: XHR.send received a load event.
PASS successfullyParsed is true

TEST COMPLETE
